refactor predicate logic out into a separate file
