+++ internal/caravan/movegoods.lua
@@ -3,6 +3,7 @@
 local common = reqscript('internal/caravan/common')
 local gui = require('gui')
 local overlay = require('plugins.overlay')
+local predicates = reqscript('internal/caravan/predicates')
 local utils = require('utils')
 local widgets = require('gui.widgets')
 
@@ -572,7 +573,7 @@ function MoveGoods:get_choices()
                 goto continue
             end
         end
-        if not common.pass_predicates(data.item, self.predicates) then
+        if not predicates.pass_predicates(data.item, self.predicates) then
             goto continue
         end
         table.insert(choices, choice)
